Add Apache2 license to the repo
Whilst the stanard Vibrato open source license is MIT this repo
contains contain in the pl_mongo library that is licensed under the
Apache License.

It doesn't make a lot of difference from a company perspective and it's
easier if the whole repo is licensed the same way so make it Apache2.

Closes: #1